Exp #11
While running exp#9 and exp#10, the moving average accuracy dropps to 9% or 2% at 16701 iteration (exp#10) and 23700 iteration (exp#9). The pattern is usually the system predicts 0. Because we have dynamic threshold, it means that all prediction becomes 0. How that possible?
I will reduce the update weight. It will take longer to converge.
I reduce the negative update weight by 0.0001. I thought that the reason I get the 0 prediction weight is that I reduce the weight too much.
I am testing it on short txt.