Exp #13
The paper uses a shortcut. Instead of tuning the permanence, they just sample from activated cells to create a dendrite and just attach it.
I don’t like this approach because
it is not biologically plausible
cells have a very focused synaptic connection to the activated cells.
The number of dendrite per cell can be changing
Instead, I will use random initialization of permanence.
We use gaussian distribution to get synapse weight. If it is less than 0 it means they are permanently not connected. About half of the cells will be permanently disconnected. We use standard deviation to be 0.15. The reason is when we calculate how many cells are connected from the beginning the following fomula gives us about 20, which looks reasonable
np.sum(np.random.randn(2042,32)*0.15 > 0.5)
stdPermanence is the design parameter we need to tune. Finally we remove the connection that is within same column